https://issues.apache.org/jira/browse/LOG4J2-3068