@carterkozak 2021-04-09
Thanks for the contribution @franz1981, your PRs are appreciated as always!

I normally handle the `changes.xml` pieces when I merge PRs since it can be non-obvious (I've certainly gotten it wrong a few times, and I don't think it makes sense to make contributors learn our system).