@franz1981 2021-04-14
@carterkozak do we have some benchmark to run this PR vs the baseline and see if it worths? 
Sorry to ask directly, but I don't have any decent HW but my laptop at hand in this very moment..